requirements.txt
google-generativeai
Pillow
python-dotenv
fastapi
uvicorn